backend/requirements.txt
fastapi>=0.115
uvicorn[standard]>=0.30
sqlalchemy[asyncio]>=2.0
aiosqlite>=0.20
alembic>=1.13
pydantic>=2.0
pydantic-settings>=2.0
jinja2>=3.1
python-dotenv>=1.0
python-multipart>=0.0.9

# LLM providers
openai>=1.0
anthropic>=0.40
google-genai>=1.0

# Search + similarity
numpy>=1.26

# Security
cryptography>=43.0
